测试了几个接口，代码微调

+++ src/main/java/com/crowdfunding/farming/mapper/OrderMapper.java
@@ -11,6 +11,6 @@ import java.util.List;
 public interface OrderMapper extends tk.mybatis.mapper.common.Mapper<Order> {
 
     List<Order> queryOrderList(
-            @Param("userId") Long userId,
+            @Param("userId") Integer userId,
             @Param("status") Integer status);
 }
